Rename uhttp/client package to uhttp/uhttpclient
"client" is not a good package name as it is far too generic.
