+++ modules/uhttp/README.md
@@ -57,7 +57,7 @@ import (
   "log"
   "net/http"
 
-  "go.uber.org/fx/modules/uhttp/client"
+  "go.uber.org/fx/modules/uhttp/uhttpclient"
   "go.uber.org/fx/service"
 )
 
@@ -68,7 +68,7 @@ func main() {
     log.Fatal("Could not initialize service: ", err)
   }
 
-  client := client.New(svc)
+  client := uhttpclient.New(svc)
   client.Get("https://www.uber.com")
 }
 ```
